some progress being able to compile changes to forward function

+++ src/lib/models/networks/DCNv2/dcn_v2.py
@@ -120,8 +120,8 @@ class DCN(DCNv2):
         o1, o2, mask = torch.chunk(out, 3, dim=1)
         offset = torch.cat((o1, o2), dim=1)
         mask = torch.sigmoid(mask)
-        return dcn_v2_conv(input, offset, mask,
-                           self.weight, self.bias,
+        return dcn_v2_conv(input.half(), offset.half(), mask.half(),
+                           self.weight.half(), self.bias.half(),
                            self.stride,
                            self.padding,
                            self.dilation,
